+++ Plone/src/elan.esd/elan/esd/content/elandoccollection.py
@@ -16,6 +16,7 @@ explanation on the statements below.
 """
 from AccessControl import ClassSecurityInfo
 from docpool.elan.config import ELAN_APP
+from docpool.event.utils import getScenarioIdsForCurrentUser
 from docpool.event.utils import getScenariosForCurrentUser
 from elan.esd import DocpoolMessageFactory as _
 from elan.esd.utils import getCategoriesForCurrentUser
@@ -69,6 +70,9 @@ class ELANDocCollection(Item, Collection):
 
     security = ClassSecurityInfo()
 
+    getScenariosForCurrentUser = getScenariosForCurrentUser
+    getScenarioIdsForCurrentUser = getScenarioIdsForCurrentUser
+
     def testSearch(self):
         """
         """
@@ -84,20 +88,6 @@ class ELANDocCollection(Item, Collection):
         # print len(res)
         for r in res:
             print(r.Title)
-
-    def getUserSelectedScenarios(self):
-        """
-        """
-        uss = getScenariosForCurrentUser(self)
-        # print usc
-        return uss
-
-    def getUserSelectedCategories(self):
-        """
-        """
-        usc = getCategoriesForCurrentUser(self)
-        # print usc
-        return usc
 
     def results(self, batch=True, b_start=0,
                 b_size=10, sort_on=None, brains=False):
@@ -225,7 +215,7 @@ class ELANDocCollection(Item, Collection):
             if not self.isArchive():
                 # First implicit filter: the user has select scenario(s) as a
                 # filter
-                uss = self.getUserSelectedScenarios()
+                uss = self.getScenarioIdsForCurrentUser()
                 if uss:
                     # This is THE modification: append the implicit criterion
                     # for the scenario(s)
@@ -248,7 +238,7 @@ class ELANDocCollection(Item, Collection):
             # Second implicit filter: the user has selected categories as a filter
             # Used for the chronological overview
             if self.isOverview():
-                usc = self.getUserSelectedCategories()
+                usc = getCategoriesForCurrentUser(self)
                 if usc:
                     value.append(
                         {